requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
idna==2.10
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.3
pytest-django==4.1.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4

django-perms~=2.1.0
timedelta~=2020.12.3